refactor(bootstrap): extract wiring output bundles to wiring/deps
Phase 4.5a of the uc-bootstrap redesign. Begin decomposing the assembly.rs
god module: move the seven wiring data types (WiringResult/WiringError,
BackgroundRuntimeDeps, SyncEngineDeps, DaemonRuntimeDeps, SharedRuntimeDeps,
WiredDependencies) into wiring/deps.rs. These carry no behavior; the wiring
logic stays in assembly for now. Contract re-exports for the deps types
source from wiring::deps; internal crate:: paths updated. assembly.rs:
1513 -> 1345 lines. No behavior change.

(deps.rs temporarily imports SystemClipboardWiring from crate::assembly;
repointed to crate::layer::platform in 4.5c.)
