+++ crates/uc-bootstrap/src/lib.rs
@@ -11,6 +11,7 @@ mod network_policy;
 pub mod observability;
 pub mod startup;
 pub mod subsystem;
+pub mod wiring;
 
 // The top-level re-exports below ARE the crate's external contract: the symbols
 // daemon (apps/daemon) and the CLI dev-tools feature (apps/cli) consume. Keep
@@ -21,10 +22,7 @@ pub mod subsystem;
 // `compose_event_context` 在 `wire_dependencies` 之后由各进程入口调用一次。
 pub use subsystem::analytics::compose_event_context;
 
-pub use assembly::{
-    get_storage_paths, wire_dependencies, BackgroundRuntimeDeps, SystemClipboardWiring,
-    WiredDependencies, WiringError, WiringResult,
-};
+pub use assembly::{get_storage_paths, wire_dependencies, SystemClipboardWiring};
 pub use entrypoint::daemon::{build_daemon_lifecycle, DaemonLifecycle};
 pub use entrypoint::non_gui::{
     build_app_facade_from_deps, build_cli_app_runtime, build_mobile_sync_facade,
@@ -35,3 +33,4 @@ pub use observability::tracing::{init_tracing_subscriber, install_panic_logging_
 pub use subsystem::blob_tasks::{spawn_blob_processing_tasks, BlobProcessingPorts};
 pub use subsystem::file_transfer::FileTransferLifecycle;
 pub use subsystem::sync_engine::SyncEngineAssembly;
+pub use wiring::deps::{BackgroundRuntimeDeps, WiredDependencies, WiringError, WiringResult};
