+++ crates/uc-bootstrap/src/wiring/mod.rs
@@ -1,0 +1,4 @@
+//! Composition-root core: the wiring output bundles and the
+//! `wire_dependencies` orchestrator that fills them.
+
+pub mod deps;
